processLog-screenshoter/requirements.txt
# requirements.txt

Flask       # app.py: 캡처 시작/중지 API 서버
mss         # app.py: 멀티 모니터 스크린샷
Pillow      # app.py: (선택적) 이미지 리사이징
schedule    # uploader.py: 주기적인 업로드 스케줄러
requests    # uploader.py: Vercel API 서버에 파일 전송